+++ app/server/models/users.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, Sequelize) => {
       allowNull: false,
       autoIncrement: true,
       primaryKey: true,
-      type: Sequelize.INTEGER
+      type: Sequelize.INTEGER(10)
     },
     username: {
       type: Sequelize.STRING(20),
